fixed hardcoded values, added tests for completed trips, refactored TravelCostService to use hashmap (simplifying direct retrieval of prices for the completed trips scenario)

+++ src/main/java/younan/george/littlepaycodingchallenge/dto/TripResult.java
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
 public class TripResult {
     private ZonedDateTime started;
     private ZonedDateTime finished;
-    private int durationSecs;
+    private long durationSecs;
     private StopId fromStopId;
     private StopId toStopId;
     private BigDecimal chargeAmount;
